Moving the components files around slightly
